fix: correct range decoder normalize and improve Opus compliance
Fix critical bug in entcode decoder normalize() where sym/rem computation
diverged from libopus — the combined sym value was stored back into rem
(growing unboundedly) instead of just the new byte, and shift ordering
differed. Also fix Encode() val offset formula to match libopus ec_encode.

Broader compliance improvements across CELT (PVQ CWRS encode/decode,
rate allocation), SILK (NLSF codebooks, LPC, decoder), opus framing
(multi-rate TOC support), and resampler tests.

+++ internal/celt/encoder.go
@@ -170,7 +170,7 @@ func (e *Encoder) Encode(samples []float64) ([]byte, error) {
 		// Write PVQ index (simplified - should use range coder)
 		// TODO: This is a simplified demo. Real Opus uses a range coder here.
 		if pulses > 0 {
-			enc.EncodeUint(uint32(pvqIndex), pulses+1)
+			enc.EncodeUint(uint32(pvqIndex), uint32(pulses+1))
 		}
 
 		// Encode fine energy
